IIT-43 Fix the PR review findings
SonarQube:
- cpp:S6012 twice, in profile_test and verify_test: explicit template
  arguments where CTAD deduces them.
- cpp:S1479 in Verify.cpp: describe() was a 32-case switch against a
  limit of 30. Now the table shape Puzzle.cpp's own describe already
  uses -- rows looked up by enumerator, a Count sentinel, and a
  static_assert on the row count. That assert only counts rows, so
  EveryViolationHasAMessage now loops to Count instead of naming
  violations by hand: a table with one enumerator twice and another
  missed is still the right SIZE, and the missed one would have
  described itself as "Unknown violation".

CodeRabbit:
- Packing.cpp: growAll's caps bound what it KEEPS, not what it has
  queued, and prepare never read the clock at all. Every expanded state
  pushes a successor per border cell, so the pending stack outgrows both
  caps by hundreds of megabytes of Bits before either trips. It now
  takes the Budget and caps the stack, and prepare records whether the
  clock or the board was what stopped it.
- Routing.cpp: dist could overflow int. One step is the PRODUCT of the
  two prices and both grow with the round, so a step reaches ~1.9e7 and
  a path of a few hundred cells sums past INT32_MAX -- undefined, and a
  wrapped distance breaks Dijkstra's only invariant. Widened to int64_t.
- logicGridSolver.css: the calc() broke stylelint's
  scss/operator-no-newline-after; the operator leads the continuation
  line now. The expression is unchanged.

ctest 7314/7314, clang-tidy gate clean, build:wasm clean, logic-grid
e2e 73 passed, and the boards the changed arms answer are unmoved.

+++ src/pages/logic-grid-solver/a-star/Routing.cpp
@@ -178,8 +178,15 @@ struct Round {
 std::vector<int> shortestPath(const Round &round, const int self,
                               const Bits &from, const Bits &to,
                               const Bits &otherTerminals) {
-  constexpr int kUnreached = INT32_MAX;
-  std::vector<int> dist(kMaxCells, kUnreached);
+  // 64-bit, and not for tidiness. One `step` below is a PRODUCT of the two
+  // prices, and both of them grow with the round: history reaches `kMaxRounds`
+  // and the pressure factor `1 + 2 * (nets - 1) * (kPressureBase + kMaxRounds
+  // / kPressureGrowth)`, so a single step reaches about 1.9e7 and a path of a
+  // few hundred cells sums past `INT32_MAX`. Signed overflow is undefined, and
+  // a wrapped negative distance would break the one invariant Dijkstra has —
+  // late rounds would re-relax cells forever without making progress.
+  constexpr int64_t kUnreached = INT64_MAX;
+  std::vector<int64_t> dist(kMaxCells, kUnreached);
   std::vector<int> prev(kMaxCells, -1);
   // A bucket queue would be tighter, but the cells are few and the costs are
   // small integers; a sorted scan of the open set is simpler to read and never
@@ -217,13 +224,14 @@ std::vector<int> shortestPath(const Round &round, const int self,
       // noise of the same size as the base step swamps the very differences
       // the price exists to express. Variation between rounds comes from the
       // order the nets are rerouted in, which is shuffled.
-      const int step =
+      const int64_t step =
           otherTerminals.test(next)
               ? kBlocked
-              : (1 + round.board.history[slot(next)] * kHistoryStep) *
+              : static_cast<int64_t>(
+                    1 + round.board.history[slot(next)] * kHistoryStep) *
                     (1 + pressureAt(round.board, next, self) *
                              round.pressureCost);
-      if (const int through = dist[slot(cell)] + step;
+      if (const int64_t through = dist[slot(cell)] + step;
           through < dist[slot(next)]) {
         dist[slot(next)] = through;
         prev[slot(next)] = cell;
